requirements/base.txt
amqp==2.4.2
billiard==3.5.0.5

certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.5
django-filter==2.2.0
django-rest-swagger==2.2.0
djangorestframework==3.9.2
idna==2.8
itypes==1.1.0
Jinja2==2.11.1
kombu==4.6.7
MarkupSafe==1.1.1
openapi-codec==1.3.2
psycopg2-binary==2.8.4
pytz==2019.3
requests==2.21.0
simplejson==3.17.0
uritemplate==3.0.1
urllib3<1.25,>=1.21.1
vine==1.3.0
easy-thumbnails==2.7

Pillow==7.0.0

django-solo==1.1.3
tqdm==4.42.1

# Vectorization
scipy==1.4.1
sklearn==0.0
keras==2.3.1
tensorflow==2.1.0
matplotlib==3.1.3
numpy==1.18.1
pandas==1.0.1
NearPy==1.0.0

celery==4.3.0
redis==3.3.8